refactored UI, began to write stuff
