+++ assets/ocr-a-extended.ttf
